1.ui完成

+++ Tools/GridMenuView.h
@@ -17,7 +17,7 @@
  * @param btntag 被选中的按钮的tag
  * @param selftag GridMenuView自己的tag
  */
-- (void)GridMenuViewButtonSelected:(int)btntag selfTag:(int)selftag;
+- (void)GridMenuViewButtonSelected:(NSInteger)btntag selfTag:(NSInteger)selftag;
 
 @end
 
@@ -26,14 +26,14 @@
     NSMutableArray *itemViewsArray;
     NSArray *btnuparray;
     NSArray *btndownarray;
-    
+
     __unsafe_unretained id<GridMenuViewDelegate> delegate;
     
     UIColor *myTitledowncolor;
     UIColor *myTitleupcolor;
 }
 
-@property(assign)int selectedbtnnum;//指示哪个按钮被选中
+@property(assign)NSInteger selectedbtnnum;//指示哪个按钮被选中
 
 @property(assign)BOOL canshowdownimg;//是否保留按钮的选中状态
 
@@ -93,6 +93,19 @@
  */
 - (id)initWithFrame:(CGRect)frame imgUpName:(NSString *)imgupname imgDownName:(NSString *)imgdownname titleArray:(NSArray *)titlearray titleDownColor:(UIColor *)titledowncolor titleUpColor:(UIColor *)titleupcolor perRowNum:(int)perrownum;
 
+/**
+ * 创建相同背景图、有title的按钮组，图片从本地获取
+ *
+ * @param frame 整个按钮组的位置和宽高
+ * @param imgupname 普通状态的按钮图片名称
+ * @param imgdownname 按下状态的按钮图片名称
+ * @param titlearray 按钮标题array
+ * @param titledowncolor 按下时按钮标题的颜色
+ * @param titleupcolor 按钮标题的颜色
+ * @param perrownum 每一行最多放几个按钮
+ *
+ * @return 按钮组view
+ */
 - (id)initWithFrame:(CGRect)frame imgUpName:(NSString *)imgupname imgDownName:(NSString *)imgdownname titleArray:(NSArray *)titlearray titleDownColor:(UIColor *)titledowncolor titleUpColor:(UIColor *)titleupcolor perRowNum:(int)perrownum andCanshowHighlight:(BOOL)showhight;
 
 @end
